requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
psycopg2
cryptography
xlrd
pandas
Pillow
